TASK: Separate adjustment configuration from base interface
Adding configuration to the base AdjustmentInterface is considered a breaking change
and would delay the introduction of this change until the next major release.
Introducing a new ConfigurationBasedAdjustmentInterface not only prevents this,
but also does not force developers to use the configuration array.
