+++ TYPO3.Media/Classes/TYPO3/Media/Domain/Model/Adjustment/AdjustmentInterface.php
@@ -18,44 +18,4 @@ use TYPO3\Flow\Annotations as Flow;
  */
 interface AdjustmentInterface
 {
-
-    /**
-     * Returns the adjustment configuration
-     *
-     * @return array
-     */
-    public function getConfiguration();
-
-    /**
-     * Sets the adjustment configuration
-     *
-     * @param array $configuration
-     */
-    public function setConfiguration(array $configuration);
-
-    /**
-     * Returns a specific adjustment configuration value
-     *
-     * @param string $path
-     * @return mixed
-     */
-    public function getConfigurationValue($path);
-
-    /**
-     * Sets a specific adjustment configuration value
-     *
-     * @param string|array $path
-     * @param mixed $value
-     * @return void
-     */
-    public function setConfigurationValue($path, $value);
-
-    /**
-     * Unsets a specific adjustment configuration value
-     *
-     * @param string|array $path
-     * @return void
-     */
-    public function unsetConfigurationValue($path);
-
 }
